Fix removing cookies in sub paths
When trying to remove a cookie in a sub path (defined via path), we can't
first try to read the cookie in a guard clause, it won't be readable
and thus the removal was not even attempted. So I'm removing the guard
clause entirely. On second thought its return value was also wrong anyway:
when attempting to remove a non-existing cookie, after the removal it still
doesn't exist and the removeCookie() should therefore return false.

Closes #336.